# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: APS - Automotive Platform System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /approvals/job/{jobId}:
    get:
      operationId: getJobApproval
      tags: [jobs]
      summary: Read the mechanic acceptance approval for a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Approval detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobApproval"
        "403": { description: Job principal or assigned mechanic required }
        "404": { description: Job or approval not found }
  /approvals/{jobId}/approve:
    post:
      operationId: approveJobApproval
      tags: [jobs]
      summary: Approve a mechanic's acceptance request
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Approval applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobApprovalDecisionResponse"
        "403": { description: Job principal required }
        "409": { description: Approval is no longer pending }
  /approvals/{jobId}/decline:
    post:
      operationId: declineJobApproval
      tags: [jobs]
      summary: Decline a mechanic's acceptance request
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeclineJobApprovalBody"
      responses:
        "200":
          description: Approval declined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobApprovalDecisionResponse"
        "403": { description: Job principal required }
        "409": { description: Approval is no longer pending }
...
  /jobs/{jobId}/lift-requirement:
    get:
      operationId: getJobLiftRequirement
      tags: [jobs]
      summary: Read the assigned mechanic's Ghost Garage lift requirement
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Lift requirement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobLiftRequirementResponse"
    patch:
      operationId: setJobLiftRequirement
      tags: [jobs]
      summary: Set the assigned mechanic's Ghost Garage lift requirement
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetJobLiftRequirementBody"
      responses:
        "200":
          description: Lift requirement updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobLiftRequirementResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    JobApproval:
      type: object
      additionalProperties: false
      properties:
        id: { type: integer }
        jobId: { type: integer }
        status: { type: string, enum: [pending, approved, declined, auto_approved, expired] }
        expiresAt: { type: string, format: date-time }
        secondsRemaining: { type: integer, minimum: 0 }
        respondedAt: { type: ["string", "null"], format: date-time }
        declineReason: { type: ["string", "null"] }
        mechanic:
          oneOf:
            - $ref: "#/components/schemas/ApprovalMechanic"
            - type: "null"
      required: [id, jobId, status, expiresAt, respondedAt, declineReason]
    ApprovalMechanic:
      type: object
      additionalProperties: false
      properties:
        id: { type: integer }
        name: { type: string }
        avatarUrl: { type: ["string", "null"] }
        mechanicTier: { type: ["string", "null"] }
        bio: { type: ["string", "null"] }
        yearsExperience: { type: ["number", "null"] }
        overallAvg: { type: number }
        reviewCount: { type: integer }
        categoriesAvg: { type: object, additionalProperties: true }
        trustScore: { type: number }
        completionRate: { type: number }
        repeatCustomerRate: { type: number }
        badges:
          type: array
          items: { type: string }
      required: [id, name, avatarUrl, mechanicTier, bio, yearsExperience, overallAvg, reviewCount, categoriesAvg, trustScore, completionRate, repeatCustomerRate, badges]
    JobApprovalDecisionResponse:
      type: object
      additionalProperties: false
      properties:
        ok: { type: boolean }
        approval: { $ref: "#/components/schemas/JobApproval" }
        jobStatus: { type: string }
      required: [ok, approval, jobStatus]
    DeclineJobApprovalBody:
      type: object
      additionalProperties: false
      properties:
        reason:
          type: string
          maxLength: 500
...
    SetJobLiftRequirementBody:
      type: object
      additionalProperties: false
      properties:
        requiresGhostGarage:
          type: boolean
      required: [requiresGhostGarage]
    JobLiftRequirementResponse:
      type: object
      additionalProperties: false
      properties:
        jobId: { type: integer }
        requiresGhostGarage: { type: boolean }
        customerTransportApproved: { type: boolean }
        status: { type: string }
      required: [jobId, requiresGhostGarage, customerTransportApproved, status]
...