# lib/api-zod/src/generated/types/jobApproval.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * APS - Automotive Platform System API
 * OpenAPI spec version: 0.1.0
 */
import type { ApprovalMechanic } from "./approvalMechanic";
import type { JobApprovalStatus } from "./jobApprovalStatus";

export interface JobApproval {
  id: number;
  jobId: number;
  status: JobApprovalStatus;
  expiresAt: Date;
  /** @minimum 0 */
  secondsRemaining?: number;
  /** @nullable */
  respondedAt: Date | null;
  /** @nullable */
  declineReason: string | null;
  mechanic?: ApprovalMechanic | null;
}